+++ packages/cap-config-writer/CHANGELOG.md
@@ -1,5 +1,49 @@
 # @sap-ux/cap-config-writer
 
+## 0.7.26
+
+### Patch Changes
+
+-   Updated dependencies [12504d5]
+    -   @sap-ux/fiori-generator-shared@0.3.14
+    -   @sap-ux/odata-service-inquirer@0.5.26
+
+## 0.7.25
+
+### Patch Changes
+
+-   Updated dependencies [df6262e]
+    -   @sap-ux/project-access@1.26.8
+    -   @sap-ux/fiori-generator-shared@0.3.13
+    -   @sap-ux/odata-service-inquirer@0.5.25
+
+## 0.7.24
+
+### Patch Changes
+
+-   @sap-ux/project-access@1.26.7
+-   @sap-ux/odata-service-inquirer@0.5.24
+-   @sap-ux/fiori-generator-shared@0.3.12
+
+## 0.7.23
+
+### Patch Changes
+
+-   Updated dependencies [9c8dc5c]
+    -   @sap-ux/odata-service-inquirer@0.5.23
+
+## 0.7.22
+
+### Patch Changes
+
+-   @sap-ux/odata-service-inquirer@0.5.22
+
+## 0.7.21
+
+### Patch Changes
+
+-   @sap-ux/odata-service-inquirer@0.5.21
+
 ## 0.7.20
 
 ### Patch Changes